requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
findspark==2.0.1
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.0
mysql-connector-python==8.0.29
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
py4j==0.10.9.5
PyInputPlus==0.2.12
pyparsing==3.0.9
PySimpleValidate==0.2.12
pyspark==3.3.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
stdiomask==0.0.6
urllib3==1.26.14
